KAFKA-6896: Add producer metrics exporting in KafkaStreams (#4998)
We would like to also export the producer metrics from StreamThread just like consumer metrics, so that we could gain more visibility of stream application. The approach is to pass in the threadProducer into the StreamThread so that we could export its metrics in dynamic.

Note that this is a pure internal change that doesn't require a KIP, and in the future we also want to export admin client metrics. A followup KIP for admin client will be created once this is merged.

Reviewers: Guozhang Wang <wangguoz@gmail.com>