KAFKA-6685: Added Exception to distinguish message Key from Value during deserializing.
https://issues.apache.org/jira/browse/KAFKA-6685

Added Exception message in `WorkerSinkTask.convertMessages` to distinguish message Key from Value during deserialization to Kafka connect format.

*More detailed description of your change,
if necessary. The PR title and PR message become
the squashed commit message, so use a separate
comment to ping reviewers.*

*Summary of testing strategy (including rationale)
for the feature or bug fix. Unit and/or integration
tests are expected for any behaviour change and
system tests should be considered for larger changes.*

Author: Jagadesh Adireddi <adireddijagadesh@gmail.com>

Reviewers: Randall Hauch <rhauch@gmail.com>, Ewen Cheslack-Postava <ewen@confluent.io>

Closes #4765 from jadireddi/KAFKA-6685---log-message-should-distinguish-key-from-value