KAFKA-7286; Avoid getting stuck loading large metadata records (#5500)
If a group metadata record size is higher than offsets.load.buffer.size, 
loading offsets and group metadata from __consumer_offsets would hang 
forever. This was due to the buffer being too small to fit any message 
bigger than the maximum configuration. This patch grows the buffer 
as needed so the large records will fit and the loading can move on.
A similar change was made to the logic for state loading in the transaction
coordinator.

Reviewers: John Roesler <vvcephei@users.noreply.github.com>, lambdaliu <lambdaliu@users.noreply.github.com>, Dhruvil Shah <dhruvil@confluent.io>, Jason Gustafson <jason@confluent.io>