MINOR: Remove throwing exception if not found from describe topics (#6112)
We recently improved the handling of the InternalTopicManager retries with #6085. The AdminClient will throw an InvalidTopicException if the topic is not found. We need to ignore that exception as when calling AdminClient#describe we may not have had a chance to create the topic yet, especially with the case of internal topics

I've created a new test asserting that when an InvalidTopicException is thrown when the topic is not found we continue on.

Reviewers: John Roesler <john@confluent.io>, Guozhang Wang <guozhang@confluent.io>