KAFKA-7288 - Make sure no bytes buffered when relying on idle timeout in channel close test (#6390)
SelectorTest.testCloseConnectionInClosingState sends and receives messages to get the channel into a state with staged receives and then waits for idle timeout to close the channel. When run with SSL, the channel may have buffered bytes that prevent the channel from being closed. Updated the test to wait until there are no buffered bytes as well.

Reviewers: Ismael Juma <ismael@juma.me.uk>