KAFKA-8442; Include ISR in Metadata response even if there is no leader (#6836)
Currently the Metadata response returns an empty ISR if there is no active leader. The behavior is inconsistent since other fields such as the replica list and offline replicas are included. This patch changes the behavior to return the current known ISR. This fixes a problem with the topic describe command which fails to report ISR when a leader is offline.

Reviewers: Jason Gustafson <jason@confluent.io>