KAFKA-8717; Reuse cached offset metadata when reading from log (#7081)
Although we currently cache offset metadata for the high watermark and last stable offset, we don't use it when reading from the log. Instead we always look it up from the index. This patch pushes fetch isolation into `Log.read` so that we are able to reuse the cached offset metadata.

Reviewers: Jun Rao <junrao@gmail.com>