MINOR: Start correlation id at 0 in SaslClientAuthenticator (#7432)
I tried to add a test for this, but it's actually pretty hard to verify what we want to
verify. I could add a test that checks the correlation field after the connection
has been established, but it would not catch this kind of bug where the issue is not
the value we store, but the value we create the request header with.

I have another PR that avoids intermediate structs during serialization/deserialization,
which has a test that fails without this change. So we'll get coverage that way.

Reviewers: Rajini Sivaram <rajinisivaram@googlemail.com>