KAFKA-9069: Flaky Test AdminClientIntegrationTest.testCreatePartitions
https://issues.apache.org/jira/browse/KAFKA-9069

Make `getTopicMetadata` in AdminClientIntegrationTest always read metadata from controller to get a consistent view.

*More detailed description of your change,
if necessary. The PR title and PR message become
the squashed commit message, so use a separate
comment to ping reviewers.*

*Summary of testing strategy (including rationale)
for the feature or bug fix. Unit and/or integration
tests are expected for any behaviour change and
system tests should be considered for larger changes.*

Author: huxihx <huxi_2b@hotmail.com>

Reviewers: Guozhang Wang <wangguoz@gmail.com>, José Armando García Sancio <jsancio@gmail.com>

Closes #7619 from huxihx/KAFKA-9069