KAFKA-7925 - Create credentials only once for sun.security.jgss.native (#6337)
When sun.security.jgss.native=true, we currently create a new server credential for every new client connection and add to the private credential set of the server's Subject. This is expensive and can result in an unbounded number of private credentials in the Subject used by the broker. The PR creates a credential immediately after login so that a single credential can be reused.

Reviewers: Guozhang Wang <wangguoz@gmail.com