KAFKA-9202: serde in ConsoleConsumer with access to headers (#7736)
The Deserializer interface has two methods, one that gives access to the headers and one that does not. ConsoleConsumer.scala only calls the latter method. It would be nice if it were to call the default method that provides header access, so that custom serde that depends on headers becomes possible.

Reviewers: Guozhang Wang <wangguoz@gmail.com>