KAFKA-9796; Ensure broker shutdown is not stuck when Acceptor is waiting on connection queue (#8448)
This commit reworks the SocketServer to always start the acceptor threads after the processor threads and to always stop the acceptor threads before the processor threads. It ensures that the acceptor shutdown is not blocked waiting on the processors to be fully shutdown by decoupling the shutdown signal and the awaiting. It also ensure that the processor threads drain its newConnection queue to unblock acceptors that may be waiting. However, the acceptors still bind during the startup, only the processing of new connections and requests is further delayed.

Reviewers: Rajini Sivaram <rajinisivaram@googlemail.com>