KAFKA-10052: Harden assertion of topic settings in Connect integration tests (#8735)
A recently added assertion in Connect integration tests uses `KafkaConsumer#partitionsFor` to verify that a topic was created with the expected number of partitions and replicas. However, probably because of metadata propagation delays, this call doesn't always return a valid `PartitionInfo` for the topic that has just been created and the test is terminated with a NPE. 

This commit changes the assertion to perform retries in order to verify the topic settings and uses the admin client to describe the topics instead.

Tests have been adjusted to use the new assertion. 

Reviewers: Randall Hauch <rhauch@gmail.com>