MINOR: Fix JSON generation of nested structs with non-matching type/name (#9277)
The schema specification allows a struct type name to differ from the field name. This works with the generated `Message` classes, but not with the generated JSON converter. The patch fixes the problem, which is that the type name is getting replaced with the field name when the struct is registered in the `StructRegistry`.

Reviewers: Colin P. McCabe <cmccabe@apache.org>