MINOR: Introduce the KIP-500 Broker lifecycle manager (#10095)
Add the KIP-500 broker lifecycle manager.  It owns the broker state.  Its inputs are
messages passed in from other parts of the broker and from the controller: requests to start
up, or shut down, for example. Its output are the broker state and various futures that can
be used to wait for broker state transitions to occur.

The lifecycle manager handles registering the broker with the controller, as described in
KIP-631. After registration is complete, it handles sending periodic broker heartbeats and
processing the responses.

Reviewers: David Arthur <mumrah@gmail.com>, Ismael Juma <ismael@juma.me.uk>, Ron Dagostino <rdagostino@confluent.io>