MINOR: Assume unclean shutdown for metadata log (#10363)
Currently the metadata log assumes successful shutdown and skips recovery. For now, we prefer to err on the safe side and assume instead that the replica was shutdown uncleanly so that we can force full recovery. This is justified in the short term because:

1. Snapshots are not fully implemented for the metadata log
2. The replicas (controllers and brokers) need to read the entire metadata log to load it into memory.

In other words, we need to read through the metadata log once on startup anyway. A long-term fix will be provided in https://issues.apache.org/jira/browse/KAFKA-12504.

Reviewers: Jason Gustafson <jason@confluent.io>