KAFKA-13053; Bump kraft frame version for incompatible changes from 2.8 (#11010)
This patch bumps the default frame version for kraft records from 0 to 1. At the same time, we reset all
records versions back to 0 and we enable flexible version support for UnregisterBrokerRecord, which was
missed previously. Note that the frame version bump also affects the KIP-405 records since they are
sharing AbstractApiMessageSerde. Since these records were not part of any previous releases, this should
not cause an issue.

Reviewers: Colin P. McCabe <cmccabe@apache.org>