KAFKA-13490: Fix createTopics and incrementalAlterConfigs for KRaft mode #11416
For CreateTopics, fix a bug where if one createTopics in a batch failed, they would all fail with
the same error code.  Make the error message for TOPIC_ALREADY_EXISTS consistent with the ZK-based
code by including the topic name.

For IncrementalAlterConfigs, before we allow topic configurations to be set, we should check that
they are valid. (This also applies to newly created topics.) IncrementalAlterConfigs should ignore
non-null payloads for DELETE operations. Previously we would return an error in these cases.
However, this is not compatible with the old ZK-based code, which ignores the payload in these
cases.

Reviewers: José Armando García Sancio <jsancio@gmail.com>, Jason Gustafson <jason@confluent.io>