MINOR: RegisterBroker should use an atomic append (#11700)
The batch of records that registers a new broker should be committed atomically. This doesn't matter
right now, because we only create a single record to register the broker. But if we create multiple
records in the future, this could matter. In order to avoid confusion, this should use
ControllerResult#atomicOf.

Reviewers: David Arthur <mumrah@gmail.com>