KAFKA-13558: NioEchoServer fails to close resources (#11618)
Due to resource leaks in the NioEchoServer, at times it won't start
properly to accept clients and will throw an exception in the
ServerSocketChannel.accept() call. Previous to this change, the error
was not being logged. The logged error was that there were too many open
files. Using the UnixOperatingSystemMXBean, I was able to detect that
use of the NioEchoServer creates several FDs but does not close them.
This then caused the client to never be able to connect to the server,
so the waitForCondition failed intermittently.

This change closes the internal Selector and the AcceptorThread's
selector so that the file descriptors are reclaimed.

Reviewers: Ismael Juma <ismael@juma.me.uk>