KAFKA-14144:; Compare AlterPartition LeaderAndIsr before fencing partition epoch (#12489)
This PR fixes an AlterPartition regression introduced in https://github.com/apache/kafka/pull/12032

When an AlterPartition request succeeds, the partition epoch gets bumped. In Zk controller mode the sender also relies on the AlterPartition response to be informed of the new partition epoch.
If the sender times out the request before a response is sent, the sender will have a stale partition epoch compared to the ZK controller state and will be fenced on subsequent AlterPartition request attempts. The sender will not receive an updated partition epoch until it receives a LeaderAndIsr request for controller-initiated ISR changes.

Reviewers: Jason Gustafson <jason@confluent.io>