KAFKA-14386; Return TopicAssignment from the ReplicaPlacer (#12892)
This changes the ReplicaPlacer interface to return a class instead of a list of list of integers. There are two reasons for the suggestion. First, as mentioned in the JIRA, it will make the interface, arguably, a bit more readable and understandable by explicitly modeling the idea of topic and partition. Second and more importantly, it makes the interface more extendable in the future. Right now it would be challenging to add more metadata to the response.

Reviewers: José Armando García Sancio <jsancio@apache.org>