KAFKA-14794; Decode base64 JSON string (#13363)
A binary value (array of bytes) can be a BinaryNode or a TextNode. When it is a BinaryNode, the method binaryValue() always returns non-null. When it is a TextNode, the method binaryValue() will return non-null if the value is a base64 string. For all other JSON nodes binaryValue() returns null.

Reviewers: Colin Patrick McCabe <cmccabe@apache.org>