KAFKA-15080; Fetcher's lag never set when partition is idle (#13843)
The PartitionFetchState's lag field is set to None when the state is created and it is updated when bytes are received for a partition. For idle partitions (newly created or not), the lag is never updated because `validBytes > 0` is never true. As a side effect, the partition is considered out-of-sync and could be incorrectly throttled.

Reviewers: Divij Vaidya <diviv@amazon.com>, Jason Gustafson <jason@confluent.io>