KAFKA-13531: Disable flaky NamedTopologyIntegrationTest (#14830)
Named topologies is a feature that is planned to be removed from AK with 4.0 and was never used via the public interface. It was used in a few versions of KSQL only, but was disabled there as well. While we do not want to remove it in 3.7 yet, we should disable flaky tests in that feature, that are disruptive to AK development.

Reviewers: Bruno Cadonna <cadonna@apache.org>