MINOR: Fix thread leak in AuthorizerIntegrationTest (#15006)
Producers and consumers could be leaked in the AuthorizerIntegrationTest. In the teardown logic, `removeAllClientAcls()` is called before calling the super teardown method. If  `removeAllClientAcls()` fails, the super method does not have a change to close the producers and consumers. Example of such failure [here](https://ci-builds.apache.org/blue/organizations/jenkins/Kafka%2Fkafka-pr/detail/PR-14925/11/tests/).

As a new cluster is created for each test anyway, calling `removeAllClientAcls()` does not seem necessary. This patch removes it.

Reviewers: Jason Gustafson <jason@confluent.io>