KAFKA-16168; Implement GroupCoordinator.onPartitionsDeleted (#15237)
This patch implements `GroupCoordinator.onPartitionsDeleted` that is called whenever a partition is deleted and must deleted all the offsets related to them. The patch uses a naive approach similar to the one used in the old coordinator. It basically iterates over all the regular end pending offsets and deletes the ones matching the deleted partition set.

Reviewers: Justine Olshan <jolshan@confluent.io>