feat: Added consumer percentage control for catch-up reading in perf test (#2573)
* Add documentation for --consumers-during-catchup parameter

This adds documentation comments for the new parameter that allows
controlling the percentage of consumers activated during catch-up reads.

* Add --consumers-during-catchup parameter to PerfConfig

Introduces a new parameter to control what percentage of consumers
should be activated during catch-up read scenarios. Default is 100%.
Includes validation to ensure the value is between 0 and 100.

* Enhance ConsumerService to support partial consumer activation

Modifies resetOffset() and resume() methods to accept a percentage parameter,
allowing only a subset of consumers to be activated during catch-up reads.
Maintains backwards compatibility with existing method signatures.

* Update PerfCommand to use consumers-during-catchup parameter

Updates the catch-up read logic to use the new parameter for controlling
the percentage of consumers that should be activated during catch-up reads.

* Fix various code style issues for whitespace and line breaks to
  comply with project standards

* fixed the two Checkstyle issues again 😭

* Implements a new --consumers-during-catchup parameter to control what percentage of topics have active consumers during catch-up reading phases. This parameter allows users to simulate realistic scenarios where only a subset of topics experience catch-up reads simultaneously.

* removes unnecessary pause